+++ CHANGELOG.md
@@ -31,6 +31,7 @@ Changed:
 - the licenes of the system is now the EUPL-1.2
 - Renamed tsm linking to data linking
 - Related units are now suggested, but all units are selectable for measured quantities
+- Harmonized field-titles "Begin description" & "End description" of the mount information
 
 Fixed:
 - SensorML validation for unit name values that contain whitespaces